Bringing worker_swarm inside kafka_avro_exporter
